omnihook: Make `cleanup` more portable
Only try to delete the standard input cache if its pathname is defined;
some older rm(1) implementations output error messages when invoked as
`rm -f` without operands [*].

  [*]: https://www.gnu.org/software/autoconf/manual/autoconf-2.71/html_node/Limitations-of-Usual-Tools.html#index-rm

+++ omnihook.m4
@@ -97,7 +97,9 @@ dnl TODO: Dynamically wrap this comment somehow.
 ifdef([USE_STDIN_CACHE],
 [# Deletes temporary files.
 cleanup() {
-    rm -f -- "$stdin_cache"
+    if test "$stdin_cache"; then
+        rm -f -- "$stdin_cache"
+    fi
 }
 
 ])dnl
